requirements.txt
datetime
json
os
python-dateutil
requests
schedule
python-dotenv
